Parameterize context path of global vars & XSpec.
As it stands, XSpec will throw errors for global `xsl:variable`,
`sch:let` and other variable bindings that make use of XPath and
a path without a context. So inside of Schematron patterns, a pattern
is set, and when running pure Schematron, it is also set for global
variables, but XSpec intentionally does not honor that assumption and
will exhibit these errors.