Update to remote resolved profile catalog. (#42)
Per conversation with @brianrufgsa, the current profiles and resolved
catalog profiles from github.com/usnistgov/oscal-content that matches
with the push from the OSCAL CI/CD system for 1.0.0-rc1. Those in the
fedramp-automation repo will be updated at a later time.

This is to further testing of the updated SSP pulled in with #41.