Sloppy manual edit, but update oscal-version.
Since 1.0.0-rc1 introduced some breaking changes, we need to check for
this now. The new upgraded example SSP does not have that field edited
since I believe it was manually created and/or manually edited after it
was run through the usnistgov/OSCAL M3->RC1 upconverter, this seems to
have been missed.