Add 'positive' role to categorize informational checks.
Per conversation with @brianrufgsa earlier this week, as we continue to
follow the dichotomy of informational validation data in sch:report and
encoding warnings and errors in sch:assert checks, it is not completely
transparent that the reports are information that is not an error, but
a useful reporting item. Brian suggested he would like this to be marked
with something like positive, so we shall start doing that and document
it.